Change Signup to Register and added useForm hook to handle user input, validation, and submission
