+++ src/components/customhooks/UseForm.js
@@ -1,22 +1,42 @@
-import React, { useState } from 'react'
+import { useState, useEffect } from 'react'
 
 
-const UseForm = (initialValues) => {
-  const [values, setValues] = useState(initialValues)
+const useForm = (formInputs, callback, validateLogin) => {
+  const [values, setValues] = useState(formInputs)
+  const [errors, setErrors] = useState({})
+  const [isSubmitting, setIsSubmitting] = useState(false)
 
-  // If initialValues.username || initialValues.passord is empty, return error?
 
 
-  return [
-    values,
-    e => {
-      setValues({
-        ...values,
-        [e.target.name]: e.target.value
-      })
+  const handleChange = e => {
+    const { name, value } = e.target
+    console.log(value)
+    setValues({
+      ...values,
+      [name]: value
+    })
+  }
+
+  const handleSubmit = e => {
+    e.preventDefault()
+    setErrors(validateLogin(values))
+    setIsSubmitting(true)
+  }
+
+  useEffect(() => {
+    if (Object.keys(errors).length === 0 && isSubmitting === true) {
+      callback()
     }
-  ]
+  }, [errors])
+
+
+  return {
+    handleChange,
+    handleSubmit,
+    values,
+    errors
+  }
 
 }
 
-export default UseForm
+export default useForm
